requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.1
dnspython==2.5.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pycparser==2.21
pymongo==4.6.1
PyMySQL==1.1.0
questionary==2.0.1